requirements.txt
pandas
streamlit
snowflake-connector-python
scikit-learn
seaborn